% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.rte.data.table}
\alias{autoplot.rte.data.table}
\title{Autplot method}
\usage{
\method{autoplot}{rte.data.table}(object)
}
\arguments{
\item{object}{A \code{rte.data.table} obtained from the API.}
}
\value{
A ggplot object.
}
\description{
Quick \code{ggplot} to explore data
}
\examples{
# todo
}
