% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eic_code.R
\docType{data}
\name{code_eic}
\alias{code_eic}
\title{EIC code table}
\format{A data.table with 158 rows and 6 variables:
\describe{
\item{code_eic}{EIC code}
\item{nom_affiche}{Id of the plant}
\item{nom_entite}{Name}
\item{eic_parent}{Parent EIC code }
\item{eic_responsible_party}{Eesponsible entity EIC code}
\item{pays}{Country}
\item{fonction}{Type of unit}
}}
\usage{
code_eic
}
\description{
This data comes from RTE, see :
\url{https://clients.rte-france.com}.
}
\examples{
code_eic
}
\keyword{datasets}
