% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-api-helpers.R
\name{get_actual_generation}
\alias{get_actual_generation}
\title{Retrieve actual generation data}
\usage{
get_actual_generation(resource = c("actual_generations_per_production_type",
  "actual_generations_per_unit", "water_reserves",
  "generation_mix_15min_time_scale"), start_date = NULL, end_date = NULL,
  token = NULL, raw = FALSE)
}
\arguments{
\item{resource}{Which ressource to use between \code{actual_generations_per_production_type}
(production data aggregated by branch), \code{actual_generations_per_unit}
(production data realised by group), \code{water_reserves}
(hydraulic stock data) and \code{generation_mix_15min_time_scale}
(production data realized from the global energy mix).}

\item{start_date}{Optional, starting date to filter results.}

\item{end_date}{Optional, ending date to filter results.}

\item{token}{Token obtained with \code{\link{get_token}}.}

\item{raw}{Return output from \code{fromJSON}.}
}
\value{
a \code{data.table}.
}
\description{
Production data aggregated by sector and group
 (in MW) on an intraday basis, corresponding to net
 production injected into the network.
}
\examples{
\dontrun{

# First you need a token
id_client <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
id_secret <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
token <- get_token(
  key = list(id_client = id_client, id_secret = id_secret)
)

# Then you can retrieve consumption data
per_prod_type <- get_actual_generation(
  "actual_generations_per_production_type", token = token
)

}
}
