% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-api-helpers.R
\name{get_consumption}
\alias{get_consumption}
\title{Retrieve consumption data}
\usage{
get_consumption(resource = c("short_term", "weekly_forecast",
  "annual_forecast"), type = c("REALISED", "ID", "D-1", "D-2"),
  start_date = NULL, end_date = NULL, token = NULL, raw = FALSE)
}
\arguments{
\item{resource}{Which ressource to use between \code{short_term}
(short-term consumption forecasts), \code{weekly_forecast}
(weekly consumption forecasts) and \code{annual_forecast}
(annual consumption forecasts).}

\item{type}{Forecast type of consumption, one or several
between 'REALISED', 'ID', 'D-1', 'D-2'.}

\item{start_date}{Optional, starting date to filter results,
if used, \code{end_date} must be set as well.}

\item{end_date}{Optional, ending date to filter results.}

\item{token}{Token obtained with \code{\link{get_token}}.}

\item{raw}{Return output from \code{fromJSON}.}
}
\value{
a \code{data.table}.
}
\description{
Data on French electricity consumption in real time.
 Forecasts are also available, they include electricity losses
 on the grid but do not include pumping consumption of waterworks.
}
\examples{
\dontrun{

# First you need a token
id_client <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
id_secret <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
token <- get_token(
  key = list(id_client = id_client, id_secret = id_secret)
)

# Then you can retrieve consumption data
consumption <- get_consumption("short_term", token = token)

}
}
