% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_api.R
\name{get_open_api}
\alias{get_open_api}
\title{Retrieve data from an open API}
\usage{
get_open_api(api, resource, start_date = NULL, end_date = NULL,
  q = list(), token = NULL, raw = FALSE)
}
\arguments{
\item{api}{Name of main API.}

\item{resource}{Resource to use in the API.}

\item{start_date}{Optional, starting date to filter results,
if used, \code{end_date} must be set as well.}

\item{end_date}{Optional, ending date to filter results.}

\item{q}{Additional parameters to query the API.}

\item{token}{Token obtained with \code{\link{get_token}}.}

\item{raw}{Return output from \code{fromJSON}.}
}
\value{
a \code{data.table}.
}
\description{
Several APIs are available, each one
 must be subscribed individually, this function allows
 to retrieve the data of an API, it is necessary for
 this purpose the name of this one and the name of the
 resource (information available on the page of the
 API on https://data.rte-france.com).
}
\examples{
\dontrun{

set_key(
  api = "physical_flow",
  key = "BASE64_KEY"
)

# Get physical flows between cross border country
physfl <- get_open_api(api = "physical_flow", resource = "physical_flows")

}
}
