% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-api-helpers.R
\name{get_physical_flows}
\alias{get_physical_flows}
\title{Retrieve physical flows data}
\usage{
get_physical_flows(
  code_eic = NULL,
  start_date = NULL,
  end_date = NULL,
  token = NULL,
  raw = FALSE
)
}
\arguments{
\item{code_eic}{Optional, EIC code of the desired country.}

\item{start_date}{Optional, starting date to filter results.}

\item{end_date}{Optional, ending date to filter results.}

\item{token}{Token obtained with \code{\link{get_token}}.}

\item{raw}{Return output from \code{fromJSON}.}
}
\value{
a \code{data.table} or a \code{list} if \code{raw = TRUE}.
}
\description{
Data about physical cross-border schedules detailing
electricity flows actually transiting across the interconnection
lines directly linking countries.
}
\examples{
\dontrun{

# First you need a token
id_client <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
id_secret <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
token <- get_token(
  key = list(id_client = id_client, id_secret = id_secret)
)

# Then you can retrieve flow data
phys_flow <- get_physical_flows(token = token)

}
}
