% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get a token to access RTE data API}
\usage{
get_token(key, user = NULL, proxy_pwd = NULL)
}
\arguments{
\item{key}{a base64 encoded string or a list containing
'client_id' and 'client_secret'. To get those credentials
you need an account on \url{https://data.rte-france.com} and
to create an application for the concerned API.}

\item{user}{Username for proxy if needeed.}

\item{proxy_pwd}{Password for proxy if needeed.}
}
\value{
a list with the access token
}
\description{
Get a token to access RTE data API
}
\examples{
\dontrun{

# To create a token you can use id_client and id_secret
id_client <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
id_secret <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
token <- get_token(
  key = list(id_client = id_client, id_secret = id_secret)
)

# or the base64 encoded key
key <- "WFhYWFgtWFhYWFgtWFhYWFgtWFhYWFgtWFhYWFg="
token <- get_token(key)

}
}
