% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pplocations.R
\docType{data}
\name{pplocations}
\alias{pplocations}
\title{Power Plants Locations}
\format{A data.table with 158 rows and 6 variables:
\describe{
\item{name}{Name of the power plant}
\item{eic_code}{EIC code}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{X,Y}{Coordinates in Lambert93}
}}
\usage{
pplocations
}
\description{
This data comes from Open-Power-System-Data, see Github repo:
\url{https://github.com/Open-Power-System-Data/conventional_power_plants}.
}
\examples{
pplocations
}
\keyword{datasets}
