% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_active_units.R
\name{retrieve_active_units}
\alias{retrieve_active_units}
\title{Retrieve generation per unit}
\usage{
retrieve_active_units(start_date = NULL, end_date = NULL)
}
\arguments{
\item{start_date}{Optional, starting date to filter results.}

\item{end_date}{Optional, ending date to filter results.}
}
\value{
a \code{data.table}
}
\description{
Retrieve generation per unit
}
\examples{
\dontrun{

library(rte.data)
library(data.table)

dat <- retrieve_active_units()
dat

dat[name \%chin\% "TRICASTIN 1"]

}
}
