% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-keys.R
\name{set_key}
\alias{set_key}
\title{Set a key for a specific API}
\usage{
set_key(api, key)
}
\arguments{
\item{api}{Name of the API.}

\item{key}{a base64 encoded string or a list containing
'client_id' and 'client_secret'. To get those credentials
you need an account on \url{https://data.rte-france.com} and
attach the concerned API to an application.}
}
\description{
Set a key for a specific API
}
\examples{
\dontrun{

set_key("consumption", "BASE64 KEY")

# or
set_key("consumption", list(id_client = "XXX", id_secret = "XXX"))

}
}
