% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy-info.R
\name{set_proxy_info}
\alias{set_proxy_info}
\title{Set proxy credentials}
\usage{
set_proxy_info(user, password, url = NULL)
}
\arguments{
\item{user}{Username for proxy.}

\item{password}{Password for proxy.}

\item{url}{Proxy url, if not provided, n attempt is made to obtain the url.}
}
\description{
Set proxy credentials
}
\note{
You'll need to restart your R session for change to be effective.
}
\examples{
\dontrun{

set_proxy_info("MYNNI", "MY_PASSWORD")
# restart your R session
# info should appear in the list
# returned by :
get_proxy_info()

}
}
