% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.rte.data.table}
\alias{autoplot.rte.data.table}
\title{Autplot method}
\usage{
\method{autoplot}{rte.data.table}(object, interactive = FALSE, ...)
}
\arguments{
\item{object}{a \code{rte.data.table} obtained from the API.}

\item{interactive}{Logical, produce an interactive visualisation.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot object.
}
\description{
Quick \code{ggplot} to explore data
}
\examples{
# todo
}
