% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_active_units.R
\name{retrieve_active_units}
\alias{retrieve_active_units}
\title{Retrieve generation per unit}
\usage{
retrieve_active_units(start_date = NULL, end_date = NULL, tokens = list())
}
\arguments{
\item{start_date}{Optional, starting date to filter results.}

\item{end_date}{Optional, ending date to filter results.}

\item{tokens}{Optional, a \code{list} of length 2 with tokens
for \code{actual_generation} and \code{generation_installed_capacities} APIs (in this order).
Can be a named list to match concerned API.}
}
\value{
a \code{data.table}
}
\description{
Retrieve generation per unit
}
\examples{
\dontrun{

library(rte.data)
library(data.table)

dat <- retrieve_active_units()
dat

dat[name \%chin\% "TRICASTIN 1"]

}
}
